shows ui components with kmp presenter, remove unused file
