+++ app-ios/Sources/ContributorFeature/SwiftUIContributorView.swift
@@ -15,30 +15,8 @@ struct SwiftUIContributorView: View {
                 ScrollView {
                     LazyVStack(spacing: 0) {
                         ForEach(contributors, id: \.id) { contributor in
-                            Button {
-                                if let urlString = contributor.profileUrl,
-                                   let url = URL(string: urlString) {
-                                    store.send(.view(.contributorButtonTapped(url)))
-                                }
-                            } label: {
-                                HStack(alignment: .center, spacing: 12) {
-                                    AsyncImage(url: URL(string: contributor.iconUrl)) {
-                                        $0.image?.resizable()
-                                    }
-                                    .frame(width: 52, height: 52)
-                                    .clipShape(Circle())
-
-                                    Text(contributor.username)
-                                        .textStyle(.bodyLarge)
-                                        .foregroundStyle(AssetColors.Surface.onSurface.swiftUIColor)
-                                        .multilineTextAlignment(.leading)
-                                        .lineLimit(2)
-
-                                    Spacer()
-                                    
-                                }
-                                .frame(maxWidth: .infinity)
-                                .padding(.init(top: 12, leading: 16, bottom: 12, trailing: 16))
+                            ContributorListItemView(contributor: contributor) { url in
+                                store.send(.view(.contributorButtonTapped(url)))
                             }
                         }
                     }
